+++ src/dipdup/config/__init__.py
@@ -934,12 +934,16 @@ class DipDupConfig:
         """
         handler_config: HandlerConfig
 
-        # NOTE: Each index must have a corresponding (currently) TzKT datasource
+        # NOTE: Each index must have a corresponding index datasource
         if isinstance(index_config.datasource, str):
-            if 'tzkt' in index_config.kind:
-                index_config.datasource = self.get_tzkt_datasource(index_config.datasource)
-            elif 'subsquid' in index_config.kind:
-                index_config.datasource = self.get_subsquid_datasource(index_config.datasource)
+            name = index_config.datasource
+            if index_config.kind.startswith('tezos.tzkt'):
+                index_config.datasource = self.get_tzkt_datasource(name)
+            elif index_config.kind.startswith('evm.subsquid'):
+                try:
+                    index_config.datasource = self.get_subsquid_datasource(name)
+                except ConfigurationError:
+                    index_config.datasource = self.get_evm_node_datasource(name)
             else:
                 raise FrameworkException(f'Unknown datasource type for index `{index_config.name}`')
 
@@ -1150,6 +1154,7 @@ def _patch_annotations(replace_table: dict[str, str]) -> None:
 _original_to_aliased = {
     'TzktDatasourceConfig': 'str | TzktDatasourceConfig',
     'SubsquidDatasourceConfig': 'str | SubsquidDatasourceConfig',
+    'SubsquidDatasourceConfig | EvmNodeDatasourceConfig': 'str | SubsquidDatasourceConfig | EvmNodeDatasourceConfig',
     'ContractConfig': 'str | ContractConfig',
     'ContractConfig | None': 'str | ContractConfig | None',
     'TezosContractConfig': 'str | TezosContractConfig',